refactor(roles): extract governance utilities to reduce LOC
CI failed: src/vibe3/roles/governance.py exceeded 400-line limit (627 lines).

Changes:
- Create governance_utils.py: extract helper functions with minimal dependencies
  - build_issue_context: issue context builder
  - build_broader_repo_entries: repo issue scanner
  - get_vibe_task_issue_numbers: vibe-task label filter
  - normalize_material_name: material name normalizer
  - find_material_in_catalog: material catalog lookup

- Create governance_factory.py: extract factory function
  - build_default_governance_fns: governance functions bundle builder
  - Avoids circular imports by importing from governance module

- Update governance.py: reduce from 627 to 384 lines
  - Import utilities from governance_utils
  - Update function references to use imported names

- Update callers:
  - scan_service.py: use find_material_in_catalog from governance_utils
  - scan.py: import build_default_governance_fns from governance_factory

Line counts:
- governance.py: 384 lines (✅ under 400)
- governance_utils.py: 205 lines (new)
- governance_factory.py: 65 lines (new)

All tests pass:
- 125 execution tests ✅
- 25 scan command tests ✅
- mypy clean ✅
- ruff clean ✅
